pyproject.toml
[project]
name = "coral-contextengineering-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai==0.126.0",
    "python-dotenv",
    "crewai-tools==0.47.1",
    "crewai-tools[mcp]",
   
]
